Move broadcast_map into comp, map_as_qty into ixmp.compat.comp
